création algo mlp sur minst

+++ src/KNN/MNISTLoader.java
@@ -33,7 +33,7 @@ public class MNISTLoader {
 
             for (int k = 0; k < nbImages; k++) {
                 images[k] = new Imagette(nbLignes, nbColonnes);
-                int[][] pixels = images[k].pixels;
+                double[][] pixels = images[k].pixels;
 
                 for (int i = 0; i < nbLignes; i++) {
                     for (int j = 0; j < nbColonnes; j++) {
@@ -56,10 +56,10 @@ public class MNISTLoader {
             pixelMap[i] = (i << 16) | (i << 8) | i;
         }
 
-        int[][] pixels = image.pixels;
+        double[][] pixels = image.pixels;
         for (int i = 0; i < nbLignes; i++) {
             for (int j = 0; j < nbColonnes; j++) {
-                bufferedImage.setRGB(j, i, pixelMap[pixels[i][j]]);
+                bufferedImage.setRGB(j, i, pixelMap[(int) pixels[i][j]]);
             }
         }
